feat: Implement OAuthClientsLimitExceeded dialog
Cozy-stack will now limit the number of registered OAuth clients based
on `cozy.oauthclients.max` flag

On web, when the limit is reached and the user tries to open a cozy-app
then the browser is redirected to `/settings/clients/limit-exceeded`

On Flagship app, we want to intercept this mechanism in order to keep
control on how application behaves

To handle this we manually do the limit check in `CozyProxyWebView`
before initializing HTML content. This would allow us to handle both
Stack and HttpServer based cozy-apps

Then we display a dedicated WebView that will display the
`/settings/clients/limit-exceeded` route from cozy-stack

Also we chose to always display cozy-home even if the limit is reached.
This is possible because cozy-home is always served from the local
HttpServer and this will ensure we don't have side effects as the
Flagship app architecture has been created around the fact that the
HomeView is always displayed

Related PR: cozy/cozy-stack#4080
Related PR: cozy/cozy-stack#4086
Related PR: cozy/cozy-stack#4087