fix: Solve blank screen when asking notification permission during webview loading
On Android, when we ask for notification permission, app goes to background. This was causing HttpServer to stop (which is an intented behavior https://github.com/futurepress/react-native-static-server/issues/39), but it was stopped during Home webview loading leading to a blank screen.

So now, we ask for notification permission before loading a webview.