feat(backup): Store more data in local storage for backed up media
We will change in a next commit the method to check if medias are
identical using these data.

We also want to be resilient to future change of our
algorithm. Especially, having local uri and remote id allow to get
information from local side and remote side quickly whenever we want.