feat: Handle App's icon in white labels
Now that we neutralized the App's main icon name, we want to use the
white labels mechanism to override this icon on MaBulle version

The base icon is already handled in each brand folder, so nothing needs
to be done on this part

Also we don't want the App's icon to change for MesPapiers in the Cozy
version anymore. This should happen only on the MaBulle version. So we
want to add the `icon/config.ts` file in each brand folder