fix: Do not set the size of the worker to 0 when hidden
We had a problem with the MGEN clisk which, when we set width and height to 0, caused the website
to loop infinitely. Because of this loop, the worker could be killed.

In order to stay as close as possible to the user experience when a clisk is launched, we decided
to set the width and height to 100%.