fix: Intercept redirection to `/settings/clients/limit-exceeded` page
In some edge-case scenario, the app may incorrectly open a cozy-app
even if the OAuth client limit is exceeded

This may happen if the `cozy.oauthclients.max` flag is not set when the
app starts but it is set just after that. In that case, the limit check
won't be done when the user tries to open a cozy-app because flags are
not updated after the app start. This is an edge case of
a8ad127b0349715db75b11b6c411df6feb2ff1fe

This may also happen if some network errors occur when querying the
cozy-stack for doing the check. In that case the app will fallback to
an "under the limit" scenario

In those cases, the app will tries to open the cozy-app and then the
cozy-stack will return a `302` to `/settings/clients/limit-exceeded`

With current implementation, this `302` will be intercepted by the
`ReloadInterceptorWebView` and redirected to an InAppBrowser instead of
the `OauthClientsLimitExceeded` popup

Note that this happens only if app is served by the cozy-stack and not
by the HttpServer

To prevent this we want to detect this scenario in the
`ReloadInterceptorWebView` and open the `OauthClientsLimitExceeded`
popup