fix: No more empty message when error before fetch
When an error occured in the preparation phase of a clisk konnector
(ensureAuthenticated, getUserDataFromWebsite), we got an empty error
message.

This is fixed, and we could have detected it with ts-check header in the
file. This will come in another PR, since there are many problems to
fix when ts-check is activated.