refactor: Support 'scanner' and 'backup'  feature in isAvailable
We decided to move from isFeatureAvailable() to
isAvailable(featureName: string) to check if a feature is available.

These feature have already been released in production. So I add the
possibility here to check if they are present by calling
isAvailable('scanner') or isAvailable('backup') but I do not remove
isScannerAvailable method and backup_available injected variable.

We will need to wait that the new isAvailable is released and
distributed to all of our users before removing the deprecated
methods.