fix: Make cozy-apps responsive to keyboard on Android
On Android and on small screen, some cozy-app fields may be hidden
behind the keyboard when it is open

This can happen on Harvest unlock form, in MesPapiers when adding some
documents, or in cozy-notes when editing a text on the bottom of a note

To prevent this we want to add a `KeyboardAvoidingView` around cozy-app
WebViews like it is already done for login, onboarding and lock screens

Also, in the addition of `KeyboardAvoidingView`, we need to remove the
`justifyContent: 'center'` on the HomeScreen. With this value the
HomeView would stay center in the screen referential when the keyboard
is open, so the HomeView would become partially hidden by the keyboard
and would have a large padding on its top

Removing `justifyContent: 'center'` on the HomeScreen seems to have no
side effect. This value is here since the beginning of the project and
it seems not to be useful anymore