feat: Use activity detection handler for motion changes
We used to rely on the activity stored in locations points to send
motion_activity to the server. However, we noticed the activity is not
always specified in the points, which was causing bad section
segmentation.

Now, we rely on the activity change API: each time an activity is
detected, we store it locally. Then, at upload time, we extract the
relevant activities (i.e. timestamp lower than the last location point,
and avoid activity duplicates), and remove it afterwards.