feat: Enable CSS vars in disableGetIndex mode
When disabling the HTTP server in local development
 environment (see `dev-config.ts`),
The CSS variables `--flagship-top-height`
and `--flagship-bottom-height` are not set
in the WebView. To counter this, we inject them manually at runtime.
This code block is never intended to be executed in production.

Still, we wrap it in a try/catch bloc to avoid breaking the WebView
in case of error.