feat: Prepare data before batch split
The batching was making the process a bit tedious,
with various edge cases.
To simplify, we now process all the locations points
and split it afterwards. 
Note a failure in the middle of a batch might cause
unexpected results, typically if the batch split was in 
the middle of a transition. However, we can also expect 
the server to handle those edge cases, but this is yet to 
be proven.
   
We also significantly increase the batch size to reduce
those risks.