feat: Filter non-heading locations after stationary activity
We noticed that sometimes, a location point will be produced after a stationary event, with a -1 heading.
This is typical of a network location and might happen
when the device enters a place with wifi.
However, we noticed that it tends to artificially extends a trip
duration. So we now filter those locations.