refactor(backup): Add metadata to uploaded file directly in upload query string
Adding metadata when uploading a file was done by creating a metadata
object and passing the metadata object id during the file upload.

I switch to pass the metadata object directly in query string. I did
not do it like this before because it was considered as a deprecated
method. After discussion with the cozy-stack team, it is now only a
not recommended method. We just need to be careful with the size of
the metadata object.

In the backup feature, we are really interested by the metadata in
query string way because it can reduce a lot the number of request
sent to the server. Especially because we will in a next commit add
an id and a creationDate in the metadata object, so every metadata
object will be different and it will not be possible to reuse
metadata object id anymore.