fix: Avoid blue screen when app is woke up by geolocation in background
react-native-background-geolocation can start the app in background. When the app was started in background, the httpserver was silently failing to start. So when the user opened the app, the app was already started but without the httpserver, and the user could see a blue screen.

Setting keepAlive httpserver parameter to true allow the server to start in background. Some considerations :
- it is not a security issue because we have the security key
- if the user switch between background and foreground, it still works
- if the app is suspended by the operating system, it still works
- if the app is killed by the operaring system, it still works