feat: Adjust default tracking frequency
Note the tracking is distance-based and not time-based.
The elasticity param is useful to reduce the frequency based
on the speed: the more fast is the device, the greater is the
minimal distance to capture a GPS point.

feat: Adapt default tracking distance

After tests, we noticed that the current tracking frequency is a
bit weak for city trips, location points being too sparsed.
Thus, we decrease the `distanceFilter` as it can be quite low as long
as the `elasticityFilter` is high enough to cope with high speed
trips, in order to preserve battery.