feat: Add STOP_TRACKING transition
This one is quite good.

Without this transition, the server was trying to attach a trip to the
end of the previous one.
And when the end point of the previous trip and the start point of the
new trip were not quite the same, unexpected behaviour were occuring.

It should fix those following:
- Wrong date for starting point, computed based on the speed between the
  previous and current trip.
- Wrong start point, that was using the end of the previous trip. This
  was typically resulting in a start point not linked to the actual
  segment in CoachCO2
- Fake trips: some trips were produced corresponding to nothing

For the explanation on server-side, see:
ee https://github.com/e-mission/e-mission-server/blob/81c4314a776eff5dee61b01f1ca16a85ee267a10/emission/analysis/intake/segmentation/restart_checking.py#L96