test: Fix open handles to improve test lisibility
We have a lot of open handles flooding the end of our test results. It seems that they are due to a React issue fixed in 18+.

Dan Abramov wrote a small package to fix the issue temporarly. I added it here manually because it has an error in the semver version check. This code should be removed when upgrading React to 18+.

https://github.com/facebook/react/issues/20756#issuecomment-780927519