feat: Handle add/remove package in white labels
For now we want to add IAP support only for Cozy's app

Adding react-native-iap package in package.json is enough to add the
"in app purchase" flag to the store. This is because the react-native
auto linking feature will automatically add the BILLING permission to
Android

To fix this and in order to have a generic solution, this commit adds
the ability to set different packages configurations for each brand

This is done in `white_label/config.json` in the `package` field.
Either add the specified package version to fix it, or add `remove` to
remove it