feat: add file thumbnail/icon on sharing
Displays a thumbnail for a file.
Will display an image thumbnail if the file is an image,
otherwise will display an icon.
 Will handle bad/empty file paths and mime types gracefully,
but if it gets to that point, there are bigger problems.
Since this is just a UI component,
it does not handle or correct errors of that nature.