fix: Do not intercept manager and scheme links in cozyAppFallbackURL
With current implementation, when clicking on a login email, if the app
is already connected to a Cozy, then the app tries to open a cozy-app
with that link

We don't want that to happen as it would produce an error and also this
is not an expected scenario

To prevent that we want to ignore links that start with a scheme or
manager links

Note that this implementation is not bullet proof. Instead of
preventing forbidden URLs we should instead hande only authorised ones.
This should be done in another commit

Also this bug highlighted another bug: when the login by email scenario
ends and when the flagship certification is automatic, then the Linking
`url` event is triggered another time (in useAppBootstrap). So this
would trigger the navigation to a cozy-app

By ignoring manager links we removed that bug's side effect but the
incorrect behavior (double event) is still present. We must fix that
bug in the future