fix: don't fetch for candidateapps at start
This is triggering unecessary rerenders elsewhere in the app
when the user doesn't has files to upload.
Instead we fetch registry if we detect presence of files.
It might be slower in file upload scenario, but it cleans up a lot
app start rerenders (file upload state is linked to localMethods)