feat: Increase log and location max retention.
Increase logMaxDays from 3 to 5 days. This was a bit short, especially
for week-ends.
Increase maxDaysToPersist from 1 to 10 days. We met a situation where a
user didn't upload any data for several days because of network
failures. When the upload did happen, all the location data older than 1
day was lost.