fix: Avoid to reset backup when we do not want to
getLocalBackupConfig is wrapped in a try / catch block to check efficiently if there is already a local backup config. But we must not add more code in the try block, because in the catch block we reset the local backup config.

Everything that must follow a "backup found" case must be outside the try block.