feat: Improve first point accuracy
When a trip starts with relative high velocity, e.g. with
a bike with no walk before , the starting point might be 
captured a bit late.
To avoid this we do the following: 
- Initialize plugin with elasticity to 1
- Force location capture after first motion activity by using
a high accuracy and a large maximumAge: this parameter
Is used to get an existing location point up to 5 min before.
From some real-life tests, this is quite useful when  the
bicycle activity is detected a bit late. The drawback is that it
can be artificially make the starting date sooner, but we 
argue this is acceptable as the user will value more an accurate location point.
- Once the data is uploaded, set back elasticity to 1