fix: Convert JSInterception console.warn to log
We still want to send the console.warn() from the flagship
app to sentry, but we don't want to send the console.warn
coming from the WebApp to sentry. It makes too much noise
and since we can't fix them from the flagship app, we don't
need to send them. They should be already sent by the wepabb
to its own sentry.