refactor: use only one source of online events
- Remove direct NetInfo use in NetStatusBoundary (use service instead)
- Add more error handling
- Use hidesplashscreen inside error screen instead of provider
- Refactor waitForOnline to accept callback as function

All this should clear up conflicting authority issues on online events
NetStatusBoundary handles app mount with the help of NetService.
NetService remains react agnostic collection of functions to be used
by various components inside the application.
Possible improvements would be to have NetStatusBoundary act as runtime
manager of online events instead of having components using netservice
directly. That would require a lot more refactor though and it is not
clear at the moment if it would bring a huge benefit.