feat: Create start point based on last end
Sometimes, the first tracked point will take some time,
both on Android and iOS.
On iOS, this is because the user must move out the geofence, while on
Android, the device takes some time to react.
The openpath server is able to use the previous place
coordinates in order to correctly start, but this is not
always reliable, in particular when a FORCE_STOP_TRACKING transition
is used: in this case, no previous start place will be used.

Therefore, we implement this behaviour here, which should
improve the start point both for Android and iOS.

Note this mechanism works for any point in the data batch.