fix: Force `FORCE_STOP_TRACKING` for large and slow gap
This transition is very useful to force a trip end and avoid
to re-use a previous place that might actually be (very) far away.
It is also useful when the device stopped for a while, and the next
location point is far away from the previous one (typically because
the geofence exit took a long time). Without this transition, the
server will continue the trip, even if the device was still for
several hours.

Note it also has drawbacks, in particular for trips with pauses,
typically by train, subway, plane, etc.
In those cases, this transition can aggressively cut the
trip, resulting in unexpected results, such as missing plane
trip, or absurd subway trips because of lack of locations.

That is why we rely on the speed to avoid such situations, and
also increase what is seen as a large temporal delta, from
30min to 60min.