feat: Detect moving still activities
It sometimes happen (at least on Android) that some "still"
activities are actually in motion.
Sometimes, those quantum activities can be produced
during a whole trip.
Therefore, we detect when this happen and change it into
"unknown" activity.
Hopefully, the openpath server will succeed to make it a
trip, which is impossible with still activities.