feat: use non-native modal where needed
Here we target screens in FullWindowOverlay;
The native modal can't be seen as it's below the app on the OS level
Also removing FullWindowOverlay in SetPinView.
The reasoning is that it was added by mistake. There is no need
for this screen to force that hard to render on top of everything