fix: Rethrow error on `fetchAppDataForSlug`
The `fetchAppDataForSlug()` is consumed as if it always return a result
so we don't want to have a catch block that returns nothing. Instead we
want it to re-throw the error

This re-throw should not change the code's behaviour as
`fetchAppDataForSlug()` is called only from `getIndexHtmlForSlug()`
that is also protected by a try..catch. The only difference is that
instead of throwing from `getIndexHtmlForSlug()` when trying to parse
`fetchAppDataForSlug()` then we throw directly from the error's origin