feat(backup): Improve dedup algorithm to work with old backup
Dedup in photo backup means that we do not upload pictures if they already exists in the Cozy.

To identify if two pictures are identical, we compare name and creation date.

For new backup, we compare with our own creation date added in the io.cozy.files metadata. It just works.

But in photo uploaded by old backup, we compare with the creation date that has been taken from EXIF and can  where the timezone could have been badly managed. So if we compare stricly date, it may not work.

So here we compare only part of the date in dedup mode by ignoring the "hour" field :
- its almost impossible to have a false identity by just ignoring the "hour" field => OK
- we can miss some identity => but we accept that our dedup is not 100% accurate