feat: Handle Android purchaseToken and prorationMode
Google allows to get multiple subscription at the same time, so when
the user upgrade or downgrade their offer, then we need to revoke the
old one. To allow this we need to provide the `purchaseToken` to the
request. This info is provided by the Cloudery using the
`purchaseToken` query string

Then we also need to specify how proration is calculated, the app does
not make this choice, instead the Cloudery provide the info through the
`prorationMode` query string