feat: Infer motion activity based on speed
It sometimes happen that the motion activity is either 'unknown' or wrongly 'still'.
Such activities are filtered by the server and cannot be used
for trip segmentation, which can lead to missing trips.
Therefore, we infer an activity based on the speed: this is
obviously not very reliable, but it's probably better than
nothing, and can eventually be corrected by the server afterwards.