feat: Implement new FlagshipUI API
The FlagshipUI is responsible to apply the correct theme to status and
navigation bars base on the displayed screen or cozy-app

Previous `setFlagshipUI` API had numerous caveats:
- It was difficult to keep track of which component was rendered on top
- It was difficult to handle components unmounting as we had to know
  which theme to apply by guessing which view would be shown after
  unmounting the components
- Previous point implied the creation of many hacks in order to have
  enough context to make a choice (i.e. hideSplashscreen, leaving
  LockScreen, handle cozy-home theme, keep track of home's konnector
  state, leaving OsReceiveScreen etc)

This new implementation is an attempt to simplify the API and to reduce
the need of hacks and side effects

It is based on new out-of-react service (`FlagshipUIService`) that
keeps track of all components that use the `useFlagshipUI()` hook

When a component registers, unregister or change its color, then the
`FlagshipUIService` updates its reference array and recompute the new
FlagshipUI to be applied

To ease the rendering order calculation, we hardcoded array of ZIndex
that every component should use when calling the `useFlagshipUI()` hook

This can be done because the application UI is simple enough so we know
which screen should be on top of others

This seems to be the only reliable way to predict screens order. We
tried to guess components positions based on React tree using different
hooks (i.e. useMemo, useEffect, useLayoutEffect, useFocus), but none of
them would allows us to keep track of screens order

We chose to put this mechanism out of React so changing the status and
navigation bars would not produce any re-rendering