feat: Adapt App and Onboarding's background colors to white labels
Now that MaBulle bootsplash is white and because the new logo is meant
to be displayed on white background we want all the transition screens
and splashscreens to have a white background

We cannot use the `primaryColor` semantic color anymore as MaBulle has
a blue one and we want a white background. So we want to create new
semantic colors for background colors on the splashscreen and
onboarding screens. This would allow us to override this color for
partners using white labels