feat(backup): Use remote album even if not  linked to current device
Before, backup created remote albums with a device id in metadata. And only remote albums with the good device id were used by a backup.

With dedup mode, we do not want anymore this mechanism because we can :
- take 100 photos with an iPhone 12 and put them in an iOS album "Holiday"
- upload 100 photos with backup and put them in a Cozy Photos album "Holiday"
- buy an iPhone 15 and migrate with Apple our iOS photos and iOS albums in the new iPhone 15
- take 10 new photos and put them in iOS album "Holiday"
- upload 10 photos with backup (previous 100 photos will not be uploaded because deduplicated)
=> even if if is not the same device id, we want to have the 110 photos in the same "Holiday" album on Cozy Photos

So now we use remote albums if they *have* a backupDeviceIds <=> if they have been created by a backup.