fix: Manage initial local notification
When the app generate a local notification, the user quit the app,
and the user start the app by clicking on the local notification, the
notification was not received by our handlers. Now it is.

notifee.getInitialNotification is noted as deprecated on notifee
documentation, but currently it is the easier solution that work both
on iOS and Android.