fix: Avoid potential duplicate initial local notification action
For notifee.getInitialNotification, we can see that it is "deprecated
for iOS in favour of onForegroundEvent, you can still
use this method on iOS but you will also receive a
onForegroundEvent". Even if we never had the case because the
onForegroundEvent
handler is called too late to catch an initial notification in our
app, we add a lastLocalNotificationId in case
we receive two times the notification.