feat: Check geolocation quota at trip end
Geolocation quota is currently only checked at startup. So if the user keep the app in background for a long time, he will be able to use geolocation for a long time even if he should not use it anymore.

We try to keep a strong isolation in the geolocation tracking service because it is used by cozy-flagship-app and CozyGPSMemory.
To be able to listen to trip end events from the geolocation service, so from the heart of CozyGPSMemory, we add an event emitter.

This event emitter will emit a trip end, which can be listened to by cozy-flagship-app.