fix: Remove cache in geolocation quota cozy-client query
**About the expected behavior**

We want to do this request until we have a timeserie. As soon as we have a timeserie, we do not need anymore to do this request.

I think this behavior is not possible to model in cozy-client.

So I do the request as long as the result is [], and then when there is a timeserie, I store the timeserie in AsyncStorage. And if there is a timeserie in AsyncStorage, I don't do the request.

**About this commit**

With the cozy-client store and 24 hours cache as before this commit, I could store [] in the cache which is NOT what I want.