fix: Do not close cozy-app when opening ClouderyOffer view
When calling `showClouderyOffer()` from a cozy-app, then this cozy-app
would be closed to display the ClouderyOffer view

This is not ideal as the `back` button would return to the Home view
instead of the previously displayed CozyApp

This was originally done because we followed the same model as the
OauthClientsLimitExceeded view. We wanted this for this view because it
was here to prevent the user to open cozy-apps. Also this view was
placed on the HomeScreen component so the user would still be able to
open the CozyApp screen displaying `cozy-settings` (so they can manage
their oauth clients)

We wrongly thought that the ClouderyOffer had the same needs, but it is
not the case. This view should be able to be displayed on top of
HomeView but also on top of CozyApp screen, and it should not close the
current CozyApp screen