fix: Do not log geolocation plugin log in Sentry
Our config only send warn and error logs to Sentry. But when an error log is sent to Sentry,
previous logs, including info logs, are also sent with it (in the "breadcrumb" field from Sentry).
This is an issue because we currently log geolocation data as info logs. So sometimes
geolocation data can be sent to Sentry if it is logged just before an error.

devlog solves the issue because it does NOT log in production mode. geolocation plugin log
will still be available in dev console and through the "Send logs" button in coachco2.