fix: Add by default the flagship UA on SupervisedWebview
We saw in our log that some requests didn't have the flagship UA. Mostly
the ones on the PasswordView:

`
2024-02-06T06:22:02.095Z  200  Mozilla  GET
https://XXX.mycozy.cloud/assets/fonts/Lato-Bold.immutable.woff2
`

It was hard for us to know whether a user had only used the flagship
app or not.

So instead of adding the userAgent on each webview, I decided to add it
to the SupervisedWebview since each webview rely on it. Like that, if
we want to override the props, we can, but by default the flagship UA
agent is set.

So I removed all the `applicationNameForUserAgent` that call the
SupervisedWebview