fix: Show SplashScreen while sending log email
On iOS, the `FileLogger.sendLogFilesByEmail()` method shows the
`MFMailComposeViewController` on top of the app UI

Because we call this when receiving an Universal Link clicked outside
of the app, this means that the Mail composer will be called when the
app gets the focus back. The result is that the Mail composer may
temporarily be displayed on top of the SplashScreen

This is problematic due to the way `react-native-bootsplash` works on
iOS

When calling `hideSplashScreen` this module will try to hide the
topmost UIViewController. This means that if the Mail composer is
displayed on top of the SplashScreen, then `hideSplashScreen` will hide
the Mail composer instead of the SplashScreen

To prevent this we want to be sure that the SplashScreen won't be
hidden until the Mail composer is closed. This can be done by
displaying a named SplashScreen before calling
`FileLogger.sendLogFilesByEmail()`, then the SplashScreen will be kept
on screen even if `hideSplashScreen` is called for the lock screen or
the global SplashScreen

In the future we may want to also fix the `react-native-bootsplash`
behaviour

Note that this problem does not occur on Android because an external
intent is called and also because the `react-native-bootsplash`
implementation is more robust on this OS