fix(upload): Save current upload id as soon as we can on iOS
We save the current upload id in a global variable to be able to
cancel an upload. On iOS, we saved it in the "begin" callback of
RNFS.uploadFiles function. We can save it sooner, before the start of the upload, so
before the "begin" callback, so let's do it to avoid having an not up
to date current upload id.