fix(backup): Canceling a backup was not properly handled on Android
When the user was canceling a backup, we rejected with an error object
where detail property was "Upload canceled" and then we checked later
if the detail property was "User cancelled upload" to do stuff related
to cancellation. It was obviously not working.

Let's fix this by creating a CancellationError extending from Error
to never reproduce this kind of issue. CancellationError is also now
a standalone error and not a subset of upload errors.