fix(backup): Do not upload a last media when clicking on the "Stop backup" button
When the user try to stop the backup by clicking on the "Stop backup"
button, it was possible that a last media was uploaded. So the user
clicked "Stop button" with "10/100 medias uploaded" displayed, and it
stopped on "11/100 medias uploaded".

When we click on the "Stop backup" button, we :
- set a "shouldStop" boolean to true
- cancel the current upload if it exists and stop backup directly

The "shouldStop" boolean was only checked at the beginning of the
upload loop. The upload loop consists in :
1. "shouldStop" check
2. metadata + dedup stuff <-- if we stop upload during this phase, we
still do the upload and we stop upload at the next iteration
3. upload stuff

In this commit, we add a "shouldStop" check between 2. and 3.