feat(backup): Stop backup immediately when going to background
Background backup is mostly unsupported. We were checking at the beginning
of the upload loop if we were in the background to stop the backup if
needed. This was highly inefficient because when going to background
the app could be freezed by the operating system before reaching
the next loop cycle. Now we listen to AppState changes to stop the
backup.

"stopBackup" can now be called with two reason leading to two error messages
sent to the user :
- stopped by the user
- stop because going to background