feat(backup): Patch react-native-cameraroll to return source type
Allow to return the source type of an asset. On iOS, the source type of an asset can be the user library (sourceType will be UserLibrary) or an iCloud Shared Album (sourceType will be CloudShared).

In the meantime, I fixed missing filename and fileSize introduced in version 7.4.1.

Upstream PR : https://github.com/react-native-cameraroll/react-native-cameraroll/pull/588