fix(geoloc): Stop tracking before deleting user data when logging out
When we stop tracking, we upload last traces, so we need user id and
user webhook. These values are stored in local storage. But when
logging out, we delete these values first, and then we try to upload
last traces. So it was creating a new id and the webhook call was
failing.