feat: bypass 2FA when logging with email
Using a react ref to store the param,
as to not disturb the existing flow.
cozy-client has been updated to use the new loginFlagship method.
Keep in mind this has not been deployed into prod stack-wise
at the time of this commit. The feature will not work with only
the front-end implementation here.