fix: Missing tracking events
We used to rely on a single `startTracking` method that was both
initializating the tracking, and started it. Furthermore, the event
handlers were made outside of this function.
This seems to be an anti-pattern, and we suspect that it was causing
tracking failures: we noticed several cases of missing
motion events, which might be the result of a wrong init.
Now, the init is done at each app start, and the tracking is actually
started when we detect the state is not enabled, just like before.