fix: Hide LOCK_SCREEN's splash screen in the correct places
With previous implementation we blindly hid the LOCK_SCREEN's splash
screen when the app would resume

This was made in #1139 to ensure we don't forget any path

But we found that this approach is too naive as it would fail when the
app requests a restart

This is because, on Android, when the app requests a restart, the
current Activity is not directly destroyed and the app react as if it
was sent to `background`. In that scenario the hook would show the
LOCK_SCREEN's splash screen and then would fail to hide it, as it would
consider the app is starting from scratch

This side effect has been fixed in previous commit as the entire
component's tree is now unmount before restart and so the LOCK_SCREEN
is not shown anymore. But we still want to revert to specificly placed
call instead of global one, so we have better control on what is
happening

Related PR: #1139