fix: Avoid start/end transitions overlap
Consider this scenario:
- A location point P1 is produced at 12:00:00 (T1), the last of a trip.
- An upload is done, with the stop transitions set at T1 + 1s =
  12:00:01
- P2 is produced at 12:00:01 (T2), the first of new trip
- At upload time, the start transitions will be set at T2 - 1s =
  12:00:00

Then, the start transition timestamp of the new trip will be before the stop
transition of the previous trip, which is not possible, and will be
discarded by the server, which will result in a missing trip.

To solve this, we now add/substract 1ms to the transitions w.r.t. to the
first or last location point. This should prevent overlapping transitions,
as long as locations between 2 trips are not produced in the same ms