feat: Dark mode in webview background
Dark mode is supported for CozyApp loading screen. But this loading screen is hidden when the webview is loaded so before the app in the webview is loaded so we could still see a white screen.

Here we set the default background color for a CozyApp webview to avoid the white screen.