feat: Allow to choose user or device color scheme
In some component, we can not use user color scheme
because we are not logged in so we can not use
the flag ui.darkmode.enabled. In this case,
getColors was defaulting to 'light'.

So here we add the ability to choose if we want to
use user or device color scheme.

Component before logged in must use device color scheme
and component after logged in (so flag present) must use
user color scheme.